refactor: extract runtime action adapter
